refactor component, create `DefaultCalendarEventRenderer`
