+++ src/index.ts
@@ -1,4 +1,6 @@
 import { Calendar } from './Calendar'
+
 export * from './Calendar'
 export * from './interfaces'
+export * from './DefaultCalendarEventRenderer'
 export default Calendar
